move iid auth logic to pkg. split pems. go embed pems
