+++ services/ctl-api/internal/app/runner-auth/service/runner_auth_aws_iid.go
@@ -2,40 +2,19 @@ package service
 
 import (
 	"context"
-	"crypto/x509"
-	"encoding/base64"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
-	"regexp"
-	"time"
 
 	"github.com/gin-gonic/gin"
-	mozpkcs7 "go.mozilla.org/pkcs7"
 	"go.uber.org/zap"
 	"gorm.io/gorm"
 
 	"github.com/nuonco/nuon/services/ctl-api/internal/app"
 	"github.com/nuonco/nuon/services/ctl-api/internal/middlewares/stderr"
+
+	runneraws "github.com/nuonco/nuon/pkg/runner/auth/aws"
 )
-
-var awsAccountIDPattern = regexp.MustCompile(`^\d{12}$`)
-
-// InstanceIdentityDocument represents the JSON document returned by
-// the EC2 IMDS at /latest/dynamic/instance-identity/document.
-type InstanceIdentityDocument struct {
-	AccountID        string    `json:"accountId"`
-	Architecture     string    `json:"architecture"`
-	AvailabilityZone string    `json:"availabilityZone"`
-	ImageID          string    `json:"imageId"`
-	InstanceID       string    `json:"instanceId"`
-	InstanceType     string    `json:"instanceType"`
-	PendingTime      time.Time `json:"pendingTime"`
-	PrivateIP        string    `json:"privateIp"`
-	Region           string    `json:"region"`
-	Version          string    `json:"version"`
-}
 
 type RunnerAuthAWSIIDRequest struct {
 	Document  string `json:"document" validate:"required"`
@@ -80,10 +59,9 @@ func (s *service) RunnerAuthAWSIID(ctx *gin.Context) {
 		return
 	}
 
-	// Parse the IID JSON
-	var iid InstanceIdentityDocument
-	if err := json.Unmarshal([]byte(req.Document), &iid); err != nil {
-		s.l.Warn("runner auth iid: failed to parse IID document", zap.Error(err))
+	iid, err := runneraws.ParseAndValidateIID(req.Document)
+	if err != nil {
+		s.l.Warn("runner auth iid: document validation failed", zap.Error(err))
 		ctx.Error(stderr.ErrAuthentication{
 			Err:         errors.New("authentication failed"),
 			Description: "invalid identity document",
@@ -92,30 +70,7 @@ func (s *service) RunnerAuthAWSIID(ctx *gin.Context) {
 		return
 	}
 
-	// Basic IID field validation
-	if iid.Region == "" || iid.AccountID == "" || iid.InstanceID == "" {
-		s.l.Warn("runner auth iid: IID missing required fields")
-		ctx.Error(stderr.ErrAuthentication{
-			Err:         errors.New("authentication failed"),
-			Description: "identity document missing required fields",
-		})
-		ctx.Abort()
-		return
-	}
-
-	if !awsAccountIDPattern.MatchString(iid.AccountID) {
-		s.l.Warn("runner auth iid: invalid account ID format",
-			zap.String("account_id", iid.AccountID))
-		ctx.Error(stderr.ErrAuthentication{
-			Err:         errors.New("authentication failed"),
-			Description: "invalid account ID format",
-		})
-		ctx.Abort()
-		return
-	}
-
-	// Verify the RSA-2048 PKCS7 signature
-	if err := s.verifyIIDSignature(iid.Region, []byte(req.Document), req.Signature); err != nil {
+	if err := runneraws.VerifyIIDSignature(s.certStore, iid.Region, []byte(req.Document), req.Signature); err != nil {
 		s.l.Warn("runner auth iid: signature verification failed",
 			zap.String("region", iid.Region),
 			zap.Error(err))
@@ -127,7 +82,6 @@ func (s *service) RunnerAuthAWSIID(ctx *gin.Context) {
 		return
 	}
 
-	// Look up the runner
 	runner, err := s.getRunnerWithGroup(ctx.Request.Context(), req.RunnerID)
 	if err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
@@ -143,7 +97,6 @@ func (s *service) RunnerAuthAWSIID(ctx *gin.Context) {
 		return
 	}
 
-	// Validate account ID against install stack outputs
 	reqCtx := ctx.Request.Context()
 	if err := s.validateRunnerAWSAccountID(reqCtx, runner, iid.AccountID); err != nil {
 		s.l.Warn("runner auth iid: account validation failed",
@@ -158,7 +111,6 @@ func (s *service) RunnerAuthAWSIID(ctx *gin.Context) {
 		return
 	}
 
-	// Create token
 	token, err := s.createRunnerToken(ctx.Request.Context(), runner.ID)
 	if err != nil {
 		s.l.Error("runner auth iid: failed to create token", zap.String("runner_id", req.RunnerID), zap.Error(err))
@@ -186,39 +138,8 @@ func (s *service) RunnerAuthAWSIID(ctx *gin.Context) {
 	})
 }
 
-// verifyIIDSignature verifies the PKCS7 signature of an instance
-// identity document using the AWS public certificate for the given region.
-// The signature from IMDS /instance-identity/rsa2048 is a PKCS7/SMIME
-// signed message.
-func (s *service) verifyIIDSignature(region string, document []byte, signatureB64 string) error {
-	cert, err := s.certStore.GetCert(region)
-	if err != nil {
-		return fmt.Errorf("no certificate for region %s: %w", region, err)
-	}
-
-	sigDER, err := base64.StdEncoding.DecodeString(signatureB64)
-	if err != nil {
-		return fmt.Errorf("failed to decode signature: %w", err)
-	}
-
-	p7, err := mozpkcs7.Parse(sigDER)
-	if err != nil {
-		return fmt.Errorf("failed to parse PKCS7 signature: %w", err)
-	}
-
-	// Use the AWS cert as the trust anchor
-	p7.Certificates = []*x509.Certificate{cert}
-
-	if err := p7.Verify(); err != nil {
-		return fmt.Errorf("PKCS7 signature verification failed: %w", err)
-	}
-
-	return nil
-}
-
 // validateRunnerAWSAccountID validates the IID account ID against the
-// install's stack outputs. This is the same check as validateRunnerAWSIdentity
-// but without the IAM role ARN check (IID doesn't provide that).
+// install's stack outputs.
 func (s *service) validateRunnerAWSAccountID(ctx context.Context, runner *app.Runner, iidAccountID string) error {
 	install, err := s.getInstallByRunnerGroup(ctx, &runner.RunnerGroup)
 	if err != nil {
